requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
streamlit==1.38.0
langchain==0.3.0
langchain-experimental==0.3.0
langchain-groq==0.2.0
groq==0.11.0
tabulate==0.9.0
streamlit-authenticator==0.3.3